requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.7
multidict==6.0.5
packaging==24.1
phonenumbers==8.13.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.0
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
yarl==1.9.4
